trade_copier/requirements.txt
python-binance
websocket-client
python-dotenv
